Create board controller
